# backend/docs/swagger.yaml
basePath: /v1
definitions:
...
  domain.PARawPayload:
    properties:
      completeness:
        type: number
      fluency:
        type: number
      phonemes:
        items:
          $ref: '#/definitions/domain.PhonemeScore'
        type: array
      prosody:
        type: number
      word_scores:
        items:
          $ref: '#/definitions/domain.WordScore'
        type: array
    type: object
  domain.PhonemeScore:
    properties:
      accuracy:
        type: number
      expected:
        type: string
      phoneme_index:
        type: integer
      said:
        description: 'NBest: what the user actually said'
        type: string
      word_index:
        type: integer
    type: object
...
  domain.Response:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
    type: object
  domain.ShadowingCompleteData:
    properties:
      avg_score:
        type: number
      completed:
        type: boolean
      passage_id:
        type: string
    type: object
...
  domain.WordScore:
    properties:
      accuracy:
        type: number
      error_type:
        description: Omission, Insertion, Mispronunciation, None
        type: string
      word:
        type: string
      word_index:
        type: integer
    type: object
...
  handler.createSessionRequest:
    properties:
      mode:
        enum:
        - word
        - sentence
        - minimal_pair
        - read_word
        - shadowing
        - exam
        type: string
      scoring_level:
        enum:
        - easy
        - medium
        - hard
        type: string
      source:
        enum:
        - recommended
        - free_choice
        - daily
        - onboarding
        type: string
    required:
    - mode
    type: object
...
  handler.ingestResultRequest:
    properties:
      audio_ref:
        type: string
      content_item_id:
        type: string
      idempotency_key:
        type: string
      minimal_pair_id:
        type: string
      pa_raw:
        $ref: '#/definitions/domain.PARawPayload'
      scoring_level:
        enum:
        - easy
        - medium
        - hard
        type: string
    required:
    - idempotency_key
    - pa_raw
    - scoring_level
    type: object
...
  handler.sentenceResultRequest:
    properties:
      audio_ref:
        type: string
      order_index:
        minimum: 0
        type: integer
      pa_raw:
        $ref: '#/definitions/domain.PARawPayload'
      scoring_level:
        enum:
        - easy
        - medium
        - hard
        type: string
    required:
    - pa_raw
    - scoring_level
    type: object
...
  service.IngestResultOutput:
    properties:
      accuracy:
        type: number
      completeness:
        type: number
      fluency:
        type: number
      item_result_id:
        type: string
      prosody:
        type: number
      trust_flag:
        type: string
    type: object
...
  service.SessionDTO:
    properties:
      ended_at:
        type: string
      id:
        type: string
      mode:
        type: string
      scoring_level:
        type: string
      source:
        type: string
      started_at:
        type: string
      summary_score:
        type: number
    type: object
  service.ShadowingProgressDTO:
    properties:
      completed:
        type: boolean
      current_sentence_index:
        type: integer
      passage_avg_score:
        type: number
      passage_id:
        type: string
      sentence_status: {}
    type: object
  service.ShadowingSentenceResult:
    properties:
      can_skip:
        type: boolean
      next_action:
        description: advance | retry | skip_available
        type: string
      passed:
        type: boolean
      score:
        type: number
    type: object
...
host: localhost:8080
info:
  contact:
    email: dev@phonara.app
    name: Phonara Team
  description: |-
    Backend API cho ứng dụng luyện phát âm tiếng Anh Phonara.
    Bao gồm: Authentication, Practice Sessions, Error Profile Engine,
    Content, Shadowing, Minimal Pairs, Subscription, Exam (Speechace), Daily Challenge.
  license:
    name: Proprietary
  title: Phonara API
  version: "1.0"
paths:
...
  /sessions:
    post:
      consumes:
      - application/json
      description: Mở session luyện tập mới. Server kiểm tra gating (quota freemium).
        Nếu scoring_level không gửi, dùng default_scoring_level của user.
      parameters:
      - description: Session params
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.createSessionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Session created
          schema:
            allOf:
            - $ref: '#/definitions/domain.Response'
            - properties:
                data:
                  $ref: '#/definitions/service.SessionDTO'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.Response'
        "402":
          description: Quota exceeded
          schema:
            $ref: '#/definitions/domain.Response'
        "422":
          description: Validation error
          schema:
            $ref: '#/definitions/domain.Response'
      security:
      - BearerAuth: []
      summary: Tạo practice session
      tags:
      - Sessions
  /sessions/{id}:
    get:
      parameters:
      - description: Session UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Session
          schema:
            allOf:
            - $ref: '#/definitions/domain.Response'
            - properties:
                data:
                  $ref: '#/definitions/service.SessionDTO'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.Response'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/domain.Response'
      security:
      - BearerAuth: []
      summary: Lấy thông tin session
      tags:
      - Sessions
  /sessions/{id}/end:
    post:
      description: Đóng session, tính summary_score = avg accuracy từ toàn bộ results.
      parameters:
      - description: Session UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Session ended
          schema:
            allOf:
            - $ref: '#/definitions/domain.Response'
            - properties:
                data:
                  $ref: '#/definitions/service.SessionDTO'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.Response'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/domain.Response'
      security:
      - BearerAuth: []
      summary: Kết thúc session
      tags:
      - Sessions
  /sessions/{id}/results:
    post:
      consumes:
      - application/json
      description: |-
        Nhận PA thô từ Azure Speech SDK. Server áp Level scoring (§3b.0a), sanity-check, lưu DB, enqueue recompute Error Profile.
        Idempotent: cùng idempotency_key sẽ không tạo result trùng (BR-QA-03).
        Recording fail (no speech/noise): trả 422, không tạo result, không trừ điểm (BR-SCORE-07).
      parameters:
      - description: Session UUID
        in: path
        name: id
        required: true
        type: string
      - description: PA raw result
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.ingestResultRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Result stored
          schema:
            allOf:
            - $ref: '#/definitions/domain.Response'
            - properties:
                data:
                  $ref: '#/definitions/service.IngestResultOutput'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.Response'
        "404":
          description: Session not found
          schema:
            $ref: '#/definitions/domain.Response'
        "422":
          description: Recording failed / validation error
          schema:
            $ref: '#/definitions/domain.Response'
      security:
      - BearerAuth: []
      summary: Nộp kết quả PA (ingestion)
      tags:
      - Sessions
  /sessions/{id}/results:batch:
    post:
      consumes:
      - application/json
      description: Nộp nhiều PA result cùng lúc khi kết nối mạng khôi phục (NFR-REL-03).
        Tối đa 50 items.
      parameters:
      - description: Session UUID
        in: path
        name: id
        required: true
        type: string
      - description: Batch results
        in: body
        name: body
        required: true
        schema:
          properties:
            results:
              items:
                $ref: '#/definitions/handler.ingestResultRequest'
              type: array
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Batch processed
          schema:
            allOf:
            - $ref: '#/definitions/domain.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/service.IngestResultOutput'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.Response'
        "422":
          description: Validation error
          schema:
            $ref: '#/definitions/domain.Response'
      security:
      - BearerAuth: []
      summary: Batch ingest results (offline sync)
      tags:
      - Sessions
  /sessions/history:
    get:
      description: Trả 50 session gần nhất. Mỗi session có summary_score và signed
        URL để replay audio (consent-gated).
      produces:
      - application/json
      responses:
        "200":
          description: Session history
          schema:
            allOf:
            - $ref: '#/definitions/domain.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/service.SessionDTO'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.Response'
      security:
      - BearerAuth: []
      summary: Lịch sử practice sessions
      tags:
      - Sessions
  /sessions/in-progress:
    get:
      description: Trả session chưa ended_at để client hiển thị "Tiếp tục 2/8" (S09b).
      produces:
      - application/json
      responses:
        "200":
          description: In-progress session (null nếu không có)
          schema:
            allOf:
            - $ref: '#/definitions/domain.Response'
            - properties:
                data:
                  $ref: '#/definitions/service.SessionDTO'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.Response'
      security:
      - BearerAuth: []
      summary: Session đang dang dở
      tags:
      - Sessions
  /shadowing/{passage_id}/complete:
    post:
      description: Mark passage completed, trả avg_score tổng kết (S26).
      parameters:
      - description: Passage UUID
        in: path
        name: passage_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Summary
          schema:
            allOf:
            - $ref: '#/definitions/domain.Response'
            - properties:
                data:
                  $ref: '#/definitions/domain.ShadowingCompleteData'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.Response'
      security:
      - BearerAuth: []
      summary: Hoàn thành passage
      tags:
      - Shadowing
  /shadowing/{passage_id}/progress:
    get:
      description: Trả current_sentence_index, sentence_status (passed/current/skipped),
        avg_score.
      parameters:
      - description: Passage UUID
        in: path
        name: passage_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Progress
          schema:
            allOf:
            - $ref: '#/definitions/domain.Response'
            - properties:
                data:
                  $ref: '#/definitions/service.ShadowingProgressDTO'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.Response'
      security:
      - BearerAuth: []
      summary: Tiến độ shadowing passage
      tags:
      - Shadowing
  /shadowing/{passage_id}/sentence-result:
    post:
      consumes:
      - application/json
      description: 'Server áp Level scoring, check ngưỡng 80% (BR-SHAD-02). Trả next_action:
        advance | retry | skip_available.'
      parameters:
      - description: Passage UUID
        in: path
        name: passage_id
        required: true
        type: string
      - description: Sentence PA result
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.sentenceResultRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Result
          schema:
            allOf:
            - $ref: '#/definitions/domain.Response'
            - properties:
                data:
                  $ref: '#/definitions/service.ShadowingSentenceResult'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.Response'
        "422":
          description: Validation error
          schema:
            $ref: '#/definitions/domain.Response'
      security:
      - BearerAuth: []
      summary: Nộp kết quả câu shadowing
      tags:
      - Shadowing
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'JWT access token. Format: "Bearer <token>"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"